boards: silabs: siwx91x: Expose real layout of the flash
The Network Coprocessor on SiWx91x owns a large part of the flash. Zephyr
is not expected to access to theses areas.

However, it is still technically possible to access these. In addition, we
prefer the DTS contains a comprehensive and transparent description of the
hardware. So update the DTS with the real partitioning of the SoC.

Reference documentation is available here[1].

[1]: https://www.silabs.com/documents/public/application-notes/
     an1416-siwx917-soc-memory-map.pdf

(cherry picked from commit 6d95573df88833a2abd7a405573083fcb6c0c4f0)

Original-Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>
GitOrigin-RevId: 6d95573df88833a2abd7a405573083fcb6c0c4f0
Cr-Build-Id: 8708352815478829489
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8708352815478829489
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Iaf27844ac2aa63003fa5b0e7a66f538040036714
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6788940
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>