kernel: Update thread_is_sliceable()
This fixes several minor items related to the priority or importance
of checks in determining whether the thread can be time sliced.

A thread that is prevented from running can not be time sliced
regardless of whether it was configured for thread-grained
time slicing or not. Nor can the idle thread be time sliced.

If the thread is configured for thread-grained time slicing, then
do not bother with the preemptible or priority threshhold checks.
This maintains the same behavior, and just optimizes the checks.

If the thread is sliceable, we may as well return the size of the
tick slice since we are checking that information anyway. Thus, a
return value of zero (0) means that the thread is not sliceable,
and a value greater than zero (0) means that it is sliceable.

(cherry picked from commit 3a0784fe22a7487e9bac144731789e7dffd087cc)

Original-Signed-off-by: Peter Mitsis <peter.mitsis@intel.com>
GitOrigin-RevId: 3a0784fe22a7487e9bac144731789e7dffd087cc
Cr-Build-Id: 8693698755917253665
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8693698755917253665
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I97918495c4f93729497fbae78f6fdd7e4c1f8e07
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7379723
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>