ci: workflows: check for manifest dnm in the manifest workflow
The current CI setup sets the manifest DNM label in a workflow and
checks it in a different workflow. The one performing the check is
configured to rerun on label changes but it's been reported few times
that the rerun logic does not seem to always run effectively and there's
been cases where the manifest has been fixed but the label has not been
removed by the automation, resulting in a stale PR.

Since the manifest has its own dedicated label, fix this race condition
by checking for that label specifically in the manifest workflow rather
than in a separate one, this means that the check is always performed
after the script that sets the label.

It also means that on manifest changes the manifest step itself will
fail rather than the PR metadata check one.

(cherry picked from commit 370e0882cbadf4f603f5d41834e34e9282546b64)

Original-Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
GitOrigin-RevId: 370e0882cbadf4f603f5d41834e34e9282546b64
Cr-Build-Id: 8716039703355817313
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8716039703355817313
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3644a49f45d0b5f6025ab97db09a34bd08f41b02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6507546
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>