net: if: Allow selecting deprecated IPv6 address as src addr
This adjust the IPv6 source address selection so that it is possible
to select deprecated IPv6 address if no better preferred address is found.

From RFC 6724 chapter 5:

   Rule 3: Avoid deprecated addresses.
   If one of the two source addresses is "preferred" and one of them is
   "deprecated" (in the RFC 4862 sense), then prefer the one that is
   "preferred".

   Rule 8: Use longest matching prefix.
   If CommonPrefixLen(SA, D) > CommonPrefixLen(SB, D), then prefer SA.
   Similarly, if CommonPrefixLen(SB, D) > CommonPrefixLen(SA, D), then
   prefer SB.

So the fix allows deprecated address to be selected if it is a better
match than the preferred one. The reasoning here is that an address with
a longer matching prefix is generally considered topologically closer to
the destination. Using such a source address can lead to more efficient
routing, as it's more likely that the source and destination are within
the same network segment or a closely related one.

(cherry picked from commit aef83fce14ce005850b6557cf04bca0ac6991501)

Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: aef83fce14ce005850b6557cf04bca0ac6991501
Cr-Build-Id: 8716098855387200097
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8716098855387200097
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I6d27c98fabb3efb64f2ae084bd535fe0847ae630
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6504820
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Jon Murphy <jpmurphy@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>