kernel: Option to allow shadow stack to be reused
It seems that, at least on tests, it's common to call k_thread_create()
on a thread multiple times. This trips a check for the CET shadow stack
- namely, set a shadow stack on a thread which already has a shadow
stack.

This patch adds a Kconfig option to allow that, iff the base address and
size of the new shadow stack are the same as before. This will trigger a
reset of the shadow stack, so it can be reused.

It may be the case that this behaviour (reusing threads) is more common
than only for tests, in which case it could make sense to change the
default - in this patch, is only true if ZTEST.

Even if being enabled by default becomes the reality, it would still
make sense to keep this option - more conscious apps could avoid the
need for the shadow stack reset code altogether.

(cherry picked from commit 10d6ef0ea5d1724934f7fd60283fc9e2dd281da0)

Original-Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>
GitOrigin-RevId: 10d6ef0ea5d1724934f7fd60283fc9e2dd281da0
Cr-Build-Id: 8704751585743389537
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704751585743389537
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Idce63931cd645bc8bbe5dfb36f54f17dec5d050f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6907924
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>