net: connection: Make it possible to update local address
Allow to update the local address on a registered connection when
rebinding.

This is needed for packet sockets, as by default packet socket
will be bound to "any" interface (interface index 0), and interface
index is part of the local address registered for packet socket.
In order to be able to explicitly bind to a specific interface later, it
needs to be possible to update the local address registered for the
connection, as we need to update the interface index, which is used
by net_conn_packet_input() for packet filtering.

(cherry picked from commit cd0b195b000e8eb7de02aa91f6bd5a3c77f4ef07)

Original-Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
GitOrigin-RevId: cd0b195b000e8eb7de02aa91f6bd5a3c77f4ef07
Cr-Build-Id: 8716039703355817313
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8716039703355817313
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3f433dd49245f1c0e43a4b21fba15e0628f85d16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6507590
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>