modules: CMSIS_6: fix path and use it for TF-M
What is the change?
Fix the path for module CMSIS_6 and use CMSIS_6 module for TF-M.

Why do we need this change?

After Zephyr updated TF-M to v2.1.0,
bb037d4469842c96f5872b271490aceb0734d965 added a local copy of CMSIS_6
to stop Zephyr's TF-M from downloading the CMSIS_6 from upstream.
The correct way would be to have CMSIS_6 as a module in Zephyr (which
we have now) and pass the path of this module to TF-M.
A fork of the upstream CMSIS_6 was added to Zephyr however,
the path in west.yml makes it a lib and not a module.
Fixing the path generates the ZEPHYR_CMSIS_6_MODULE_DIR symbol which
can now be used to pass to TF-M and the copy in TF-M would no longer be
required.

(cherry picked from commit c4235fa48908737f0055c8ab883b28d0599f5b89)

Original-Signed-off-by: Sudan Landge <sudan.landge@arm.com>
GitOrigin-RevId: c4235fa48908737f0055c8ab883b28d0599f5b89
Cr-Build-Id: 8714898446203671153
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8714898446203671153
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ifc933b0d4014d9e90d2800af697aa067929dec59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6546341
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>