tests: Bluetooth: Split CAP and GMAP AC tests into separate scripts
This commit splits the CAP and GMAP audio configuration (AC)
tests into separate scripts. The purpose of this is to reduce
the runtime of the individual scripts (some, like the cap_ac_11,
tests takes a long time to run all combinations).
This split also improves support for running them in parallel.

An additional, positive, side effect of this is that the logs will
also be smaller per run.

The new scripts can be run directly, or via the e.g. _cap_ac_X.sh
scripts that run entire audio configuration for all presets.

The design of how the parameters are passed around, are based
on variable instead of function arguments. Effectively they can
be used for the same thing, but variables have the advantage
that it is easier to deal with longer names,
and that any of the test scripts can be called with additional
arguments that can, when using variables, easily be forwarded
to the call to Execute for the PHY in the
_ac_common.sh.

(cherry picked from commit 60e9f9958297fcf9eb55b30b62af247738500ab3)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 60e9f9958297fcf9eb55b30b62af247738500ab3
Cr-Build-Id: 8713584789952251569
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8713584789952251569
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I4800956a03576d6235764625e5d0c304a15c82e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6599798
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>