drivers: uart_mcux_flexcomm: Enable as wakeup src
Add the feature for the flexcomm uart to be able to be a wakeup source
from low power modes. To be able to do this on a relevant platform, the
DT node for the UART needs to have the wakeup-source property and define
a "sleep" clock. The details of handling the sleep clock and default
clock are still platform specific but handled by the clock control
driver of the platform, so that this code should be able to be
applicable to any platform that follows the same DT definition.

(cherry picked from commit e5a35e64354dad24f457459b2e27b97b9b85ba42)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: e5a35e64354dad24f457459b2e27b97b9b85ba42
Cr-Build-Id: 8713403596655375841
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8713403596655375841
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I63301408ab7d0606a75493bf6731a3e05ef10531
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6608689
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>