gen_edt: use workspace dir as base for relative paths in comments
This commit allows comments to reference files with paths that are relative
to the Zephyr workspace root. This is done by adding a new argument
'--workspace-dir' to the 'gen_edt.py' script, which is passed to the
'EDT' and 'DT' classes and used instead of the current working directory.

The workspace directory is set to WEST_TOPDIR if West is in use,
otherwise it is set to the parent directory of ZEPHYR_BASE so that
Zephyr files have a 'zephyr/' prefix.

(cherry picked from commit c1603b31634a8807c1dec31125e1ee42b7fe68f6)

Original-Signed-off-by: Luca Burelli <l.burelli@arduino.cc>
GitOrigin-RevId: c1603b31634a8807c1dec31125e1ee42b7fe68f6
Cr-Build-Id: 8714694602652640193
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8714694602652640193
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ie196719d22aa2b41b3bb8d6959a97636dae463f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6558080
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>