boards: lilygo: ttgo_tbeam: initialize GNSS after regulator
As the initialization order is not yet defined by the device tree, we
manually need to reflect the order via the init priorities. As the GNSS
is connected to a UART that is powered by a regulator, the regulator
needs to be initialized first. By setting the prio to 87, we initialize
GNSS right after the regulator (instead of before as done without this
patch).

(cherry picked from commit c35bb0de8023750de5ba31dcfbd7b5277466147e)

Original-Signed-off-by: Felix Moessbauer <felix.moessbauer@gmail.com>
GitOrigin-RevId: c35bb0de8023750de5ba31dcfbd7b5277466147e
Cr-Build-Id: 8715260833202505873
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8715260833202505873
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3ae98ef94a9e11501ac4e22add8601f694fd19a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6535165
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>