dts: st: h7: move dma property of dcmi in stm32h7.dtsi
Usage of dma is mandatory for the dcmi and this property is
tightly coupled with the soc itself since the configuration of
the dma depends on the source/destination, and the request line
is also fixed for an ip.
Instead of having to always have the dma property part of the
board or shield dts/overlay, add the dma property into the
dcmi node of the stm32h7.dtsi.

(cherry picked from commit 99e12cbf4ad1c48133f70625f1d42ff8025f588b)

Original-Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
GitOrigin-RevId: 99e12cbf4ad1c48133f70625f1d42ff8025f588b
Cr-Build-Id: 8714921172849795585
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8714921172849795585
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Id9ceb9cb3125ddd595b8e337f516c8351ad86525
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6543903
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>