soc: rw: Switch main clock on PM2 to LPOSC
On RW, normal configuration has all clock generators gated in PM2.
Only the LPOSC is available for main clock source since it is a low
power clock.

Many of the peripherals on the chip are still "on" and do need a
main clock source in order to be effective as wakeup sources
to the chip as intended. So we should make this switch for PM2
specifically in order to achieve desired wakeup capabilities.

(cherry picked from commit 4598c1875406fa6128eb087ce1c248344a5a37a6)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 4598c1875406fa6128eb087ce1c248344a5a37a6
Cr-Build-Id: 8713403596655375841
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8713403596655375841
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I944e8c623668ea3c646f858fded5dfd4e4dedfac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6608688
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>