debug: ASSERT_VERBOSE depends on PRINTK
When both PRINTK and ASSERT are disabled, the firmware image will still
contain a symbol for assert_print function which increases the memory
footprint by 14 bytes. In this case the function doesn't do anything and
so these 14 bytes are not useful.

This change also reduces the memory footprint elsewhere such as in the
bluetooth subsystem which implements BT_ASSERT_VERBOSE. Now it will be
nop. On nRF52832 this is on the order of 300 - 400 bytes when configured
for peripheral only (just as an example).

(cherry picked from commit f32223aac505b8d9b1ac962336c4e4d62181c810)

Original-Signed-off-by: Guðni Már Gilbert <gudni.m.g@gmail.com>
GitOrigin-RevId: f32223aac505b8d9b1ac962336c4e4d62181c810
Cr-Build-Id: 8713516842192680513
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8713516842192680513
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I6cfaaf8d1434b62c15a59bccd21f243fd6868bfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6602737
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>