samples: Bluetooth: BAP: Unicast Server: Refactor TX
Refactors the TX of the unicast server. This commit effectively
copies the TX approach of the unicast client sample, to make
the sample more similar.
This changes the TX to be based on the number of complete
packets event (which results in free'ing the buffer), instead of
a timer. This completely ensures that we do not skip any sequence
numbers which may be rejected by the controller.

(cherry picked from commit 27b2d17ad1d3c72d4092717cfd6667013924cddb)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 27b2d17ad1d3c72d4092717cfd6667013924cddb
Cr-Build-Id: 8704751585743389537
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704751585743389537
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I2b976ea33fd2a4b410a39b2289da408bd7cd329d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6907976
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>