xtensa: fix num_high_regs calculation when dumping stack
The calculation of number of high registers is not entirely
correct. We need to get past the pointer to BSA in the stack
frame before reaching the high registers. The location address
difference between the BSA and start of high registers then
can be used to calculate how many high registers in the stack
frame. So correct the start location of high registers in
the calculation as it was incorrect before. Though the result
would be the same as further divisions would mask this error.
However, it is better to correct this for readability.

(cherry picked from commit d00095c4a6b5c0c139c9b76878f181918db4d1c1)

Original-Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: d00095c4a6b5c0c139c9b76878f181918db4d1c1
Cr-Build-Id: 8711727552213157905
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8711727552213157905
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: If5fecc2053a77f689c03035e6fa772729ee48e28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6654764
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>