sys: clock: additional sys_clock api calls
Additional entries for the sys_clock API, comprised of:

* sys_clock_gettime()
* sys_clock_settime()
* sys_clock_nanosleep()

along with the constants

* SYS_CLOCK_REALTIME
* SYS_CLOCK_MONOTONIC
* SYS_TIMER_ABSTIME

The primary motivation for this API is so that libc and other
libraries have a familiar-enough API to reach to when POSIX
is not available, since POSIX is optional in Zephyr.

By adding this API to lib/os, we also eliminate dependency cycles
between libc and posix, as lib/os is a mutual dependency.

(cherry picked from commit a3e934f12c2a5a9b72878649ceefc4fbfeaa1e1a)

Original-Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>
GitOrigin-RevId: a3e934f12c2a5a9b72878649ceefc4fbfeaa1e1a
Cr-Build-Id: 8710934827903433057
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8710934827903433057
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0942cc141a0f77809d9881cf9c28807f73b43374
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6683240
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>