drivers: flash: add andes qspi xip flash driver
Add a flash driver that is used to perform flash operations on a flash
chip that is connected to an Andes QSPI controller and is used for XIP
mode.

The driver is as small as possible, because necessary code has to be
placed in RAM. It is not possible to fetch code from flash while
performing erase/write operations.

(cherry picked from commit 6d5cdbb13e79e0556cb8b5ac6d0efe3f3c70e010)

Original-Signed-off-by: Dawid Niedzwiecki <dawidn@google.com>
GitOrigin-RevId: 6d5cdbb13e79e0556cb8b5ac6d0efe3f3c70e010
Cr-Build-Id: 8708556658338805137
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8708556658338805137
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I8a5b829daa738993d13ff5a52d96c2227d068045
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6781261
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>