pm: device: introduce pm_device_driver_deinit()
Introduce pm_device_driver_deinit() which is required to fully
support device_deinit(). This function shall be called by drivers
when device_deinit() is called.

If PM_DEVICE is enabled, it will
simply check whether the device is either SUSPENDED or OFF, this
will prevent device_deinit() on a device which is currently in
use, and ensure the device is in the correct state if device_init()
is called later.

If PM_DEVICE is not enabled, it will invoke the SUSPEND action which
mirrors the pm_device_driver_init() behavior. Note that TURN_OFF
action is not called since the device is deinitialized after this
call.

(cherry picked from commit cd16d17dca15ab7103494e5956a004ca1dabb40c)

Original-Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>
GitOrigin-RevId: cd16d17dca15ab7103494e5956a004ca1dabb40c
Cr-Build-Id: 8707945128060306049
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8707945128060306049
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Id8dd2899c28d2d19c5278cd9cc7211053d6fece7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6799012
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>