include: usb_ch9: ensure packing of usb_setup_packet
struct usb_setup_packet represents the setup data payload of a USB control
message which is exactly 8-bytes in length.

On a legacy build of GCC 4.1.2 for an OpenRISC-like architecture it was
found that the compiler added padding after the bmRequestType anonymous
union, and subsequently the bRequest field causing the structure to be
padded to 12-bytes in length with the fields incorrectly laid out.

This patch corrects the issue by applying the __packed attribute to the
anonymous union. This corrects the issue.

The __packed attribute is also applied to the outer structure for the sake
of consistency.

(cherry picked from commit 8b5fe624e2c15e4d32036176d93bf25ffb771375)

Original-Signed-off-by: Joel Holdsworth <jholdsworth@nvidia.com>
GitOrigin-RevId: 8b5fe624e2c15e4d32036176d93bf25ffb771375
Cr-Build-Id: 8714128372271872209
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8714128372271872209
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I1f90fbec9bde0a086550170b5380081f37fa4e5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6581526
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>