FROMPULL: i2c: npcx: Special handling for erronious isr error detection
https://github.com/zephyrproject-rtos/zephyr/pull/90241

We saw an issue with a warm reboot of the npcx9m7fb where the interrupt
called even though there was no transaction. This is fixed by checking
if the 'port' field was ever set. If it wasn't then this interrupt
should be ignored as it's not a result of any action performed by the
application.

Change-Id: I69c1d1e59ef997a074f0e1e7913007e88b06f47f
Signed-off-by: Yuval Peress <peress@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6576538
Reviewed-by: Keith Short <keithshort@chromium.org>