drivers: gpio: Fix GPIO initial status
The current procedure to initialize the GPIO is:
If the GPIO is configured as GPIO_OUTPUT and also has the
GPIO_OUTPUT_INIT_HIGH flag, set it to output high; otherwise, set it to
output low.

This may fail if the GPIO is simply configured as GPIO_OUTPUT without
specifying either INIT_HIGH or INIT_LOW, which means it is intended to
preserve the previous status. But in this case, we are currently setting it
to output low.

Fix this by explicitly initializing the GPIO to high or low according to
the configuration:
If the GPIO is configured as GPIO_OUTPUT and also with the
GPIO_OUTPUT_INIT_HIGH flag, set it to output high.
If the GPIO is configured as GPIO_OUTPUT and also with the
GPIO_OUTPUT_INIT_LOW flag, set it to output low.
If the GPIO is configured as GPIO_OUTPUT only,
do not set the output value.

(cherry picked from commit 728a018f867dc88e516083da7a9126914608b264)

Original-Signed-off-by: Benson Huang <benson7633769@gmail.com>
GitOrigin-RevId: 728a018f867dc88e516083da7a9126914608b264
Cr-Build-Id: 8714196319069227041
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8714196319069227041
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0acd761334da3c12f7a529c4042b4d50987468f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6577896
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>