code_relocation: fix ruff B028 warnings
The ruff python linter produces a B028 warning when the warnings.warn()
function is called with a stacklevel parameter.

By default the function will set a stacklevel of 1 which causes it to
output the stack frame of the line where the function is called without
any context information from higher up the stack.

It is recommended to use a stacklevel of 2 or higher. Therfore, this patch
sets the stacklevel parameter of all warnings.warn() calls to to 2.

(cherry picked from commit 4f0b402224f534e102f79ccb7612a302c7b3de87)

Original-Signed-off-by: Joel Holdsworth <jholdsworth@nvidia.com>
GitOrigin-RevId: 4f0b402224f534e102f79ccb7612a302c7b3de87
Cr-Build-Id: 8714128372271872209
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8714128372271872209
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: If871ddff1fecf6581dca92cdfb4793095aa225d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6581516
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>