drivers: dma: sdma: Fix noise issue with pause/resume
Each time we configure an SDMA channel we also compute the total
allocated DMA buffer length but we assume is initialized with zero.

This is true each time a channel is requested. But if there are
multiple calls to configure without releasing the channel the buffer
length is not correctly computed.

So, we need to initialize it with zero each time we reconfigure the dma
channel.

(cherry picked from commit 312ff1c904f0afaccda84fec6dace135788e50bc)

Original-Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>
GitOrigin-RevId: 312ff1c904f0afaccda84fec6dace135788e50bc
Cr-Build-Id: 8713584789952251569
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8713584789952251569
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I6944837b79542bd1b8ad5c12228ef7aaefad82dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6599956
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>